requirements2.txt
click==8.1.7
cmake==3.30.5
colorama==0.4.6
cv==1.0.0
cv2box==0.5.9
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.26.4
opencv-python==4.10.0.84
pandas==2.2.3
pillow==11.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
tqdm==4.66.5
tzdata==2024.2